+++ apps/desktop-wallet/src/api/apiDataHooks/market/useFetchTokenPrices.ts
