+++ apps/desktop-wallet/src/api/apiDataHooks/tokens/useFetchToken.ts
@@ -21,8 +21,8 @@ import { explorer } from '@alephium/web3'
 import { useQuery } from '@tanstack/react-query'
 
 import { DataHook } from '@/api/apiDataHooks/apiDataHooksTypes'
-import useFetchFtList from '@/api/apiDataHooks/useFetchFtList'
-import useFetchNft from '@/api/apiDataHooks/useFetchNft'
+import useFetchFtList from '@/api/apiDataHooks/tokens/useFetchFtList'
+import useFetchNft from '@/api/apiDataHooks/tokens/useFetchNft'
 import { fungibleTokenMetadataQuery, tokenTypeQuery } from '@/api/queries/tokenQueries'
 import { ListedFT, NonStandardToken, TokenId, UnlistedFT } from '@/types/tokens'
