+++ apps/desktop-wallet/src/components/Inputs/SelectOptionWalletToken.tsx
@@ -16,7 +16,7 @@ You should have received a copy of the GNU Lesser General Public License
 along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
 
-import useFetchToken, { isNFT } from '@/api/apiDataHooks/useFetchToken'
+import useFetchToken, { isNFT } from '@/api/apiDataHooks/tokens/useFetchToken'
 import useFetchWalletSingleTokenBalances from '@/api/apiDataHooks/wallet/useFetchWalletSingleTokenBalances'
 import SelectOptionToken, { SelectOptionTokenBaseProps } from '@/components/Inputs/SelectOptionToken'
