+++ apps/desktop-wallet/src/components/TokenBadge.tsx
@@ -18,7 +18,7 @@ along with the library. If not, see <http://www.gnu.org/licenses/>.
 
 import styled, { css } from 'styled-components'
 
-import useFetchToken, { isFT, isNFT } from '@/api/apiDataHooks/useFetchToken'
+import useFetchToken, { isFT, isNFT } from '@/api/apiDataHooks/tokens/useFetchToken'
 import Amount from '@/components/Amount'
 import AssetLogo from '@/components/AssetLogo'
 import SkeletonLoader from '@/components/SkeletonLoader'
