+++ apps/desktop-wallet/src/features/assetsLists/tokenBalanceRow/FTWorth.tsx
@@ -19,8 +19,8 @@ along with the library. If not, see <http://www.gnu.org/licenses/>.
 import { calculateAmountWorth } from '@alephium/shared'
 import styled from 'styled-components'
 
-import useFetchToken, { isFT } from '@/api/apiDataHooks/useFetchToken'
-import useFetchTokenPrices, { useFetchTokenPrice } from '@/api/apiDataHooks/useFetchTokenPrices'
+import useFetchToken, { isFT } from '@/api/apiDataHooks/tokens/useFetchToken'
+import useFetchTokenPrices, { useFetchTokenPrice } from '@/api/apiDataHooks/market/useFetchTokenPrices'
 import Amount from '@/components/Amount'
 import SkeletonLoader from '@/components/SkeletonLoader'
 import { TokenBalancesRowBaseProps } from '@/features/assetsLists/tokenBalanceRow/types'
